Renaming connect4 algo to avoid confusion
Before the connect4 algorithm was installed under the name 'connect4'
which could create conflict with another python library.
To avoid this, the algo is now installed under 'connect4_alg'
and at the import, it will look for the local module 'connect4_alg' first (where the binary file is now imported) instead of searching at the top level.